+++ jsonschema-generator/src/test/java/com/github/victools/jsonschema/generator/FieldScopeTest.java
@@ -49,7 +49,13 @@ public class FieldScopeTest extends AbstractTypeAwareTest {
             Arguments.of("fieldWithPrivateGetter", null, null),
             Arguments.of("fieldWithPublicGetter", null, "getFieldWithPublicGetter"),
             Arguments.of("fieldWithPublicGetter", "fieldWithoutGetter", "getFieldWithPublicGetter"),
-            Arguments.of("fieldWithPublicBooleanGetter", null, "isFieldWithPublicBooleanGetter")
+            Arguments.of("fieldWithPublicBooleanGetter", null, "isFieldWithPublicBooleanGetter"),
+            Arguments.of("isFieldWithMatchingGetter", null, "isFieldWithMatchingGetter"),
+            Arguments.of("isFieldWithMatchingGetter", "fieldWithPublicGetter", "isFieldWithMatchingGetter"),
+            Arguments.of("aFieldWithJavaBeansConformingGetter", null, "getaFieldWithJavaBeansConformingGetter"),
+            Arguments.of("aFieldWithJavaBeansNonConformingGetter", null, "getAFieldWithJavaBeansNonConformingGetter"),
+            Arguments.of("aFieldWithJavaBeansConformingBooleanGetter", null, "isaFieldWithJavaBeansConformingBooleanGetter"),
+            Arguments.of("aFieldWithJavaBeansNonConformingBooleanGetter", null, "isAFieldWithJavaBeansNonConformingBooleanGetter")
         );
     }
 
@@ -107,6 +113,11 @@ public class FieldScopeTest extends AbstractTypeAwareTest {
         private int fieldWithPrivateGetter;
         private long fieldWithPublicGetter;
         private boolean fieldWithPublicBooleanGetter;
+        private boolean isFieldWithMatchingGetter;
+        private int aFieldWithJavaBeansConformingGetter;
+        private long aFieldWithJavaBeansNonConformingGetter;
+        private boolean aFieldWithJavaBeansConformingBooleanGetter;
+        private boolean aFieldWithJavaBeansNonConformingBooleanGetter;
 
         private int getFieldWithPrivateGetter() {
             return this.fieldWithPrivateGetter;
@@ -119,6 +130,26 @@ public class FieldScopeTest extends AbstractTypeAwareTest {
         @TestAnnotation
         public boolean isFieldWithPublicBooleanGetter() {
             return this.fieldWithPublicBooleanGetter;
+        }
+
+        public boolean isFieldWithMatchingGetter() {
+            return this.isFieldWithMatchingGetter;
+        }
+
+        public int getaFieldWithJavaBeansConformingGetter() {
+            return this.aFieldWithJavaBeansConformingGetter;
+        }
+
+        public long getAFieldWithJavaBeansNonConformingGetter() {
+            return this.aFieldWithJavaBeansNonConformingGetter;
+        }
+
+        public boolean isaFieldWithJavaBeansConformingBooleanGetter() {
+            return this.aFieldWithJavaBeansConformingBooleanGetter;
+        }
+
+        public boolean isAFieldWithJavaBeansNonConformingBooleanGetter() {
+            return this.aFieldWithJavaBeansNonConformingBooleanGetter;
         }
     }